move lookup full version to cros.rs
