+++ src/cmd/flash.rs
@@ -10,11 +10,10 @@ use anyhow::anyhow;
 use anyhow::Result;
 use argh::FromArgs;
 use lium::cros::ensure_testing_rsa_is_there;
+use lium::cros::lookup_full_version;
 use lium::dut::DutInfo;
 use lium::repo::get_repo_dir;
 use regex::Regex;
-
-use crate::lookup_full_version;
 
 /// Determine a BOARD to flash, based on the parameters.
 /// If arg_dut is specified, this function will check if the board given via
